requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
